feat: Implement post boosting payment flow and analytics dashboard
This commit introduces the end-to-end flow for boosting a post, including payment integration via PayU. It also adds a new Analytics Dashboard to the Professional Dashboard for detailed performance tracking.

### Key Changes:

-   **Post Boosting & Payment Flow (`CreateNewAd.java`)**:
    -   Integrated the `initiateBoostPost` API to start the boosting process.
    -   Implemented the complete PayU payment gateway flow, including hash generation, payment initialization, and handling success/failure callbacks.
    -   A new `SuccessfulWithOkay` dialog was added to redirect the user to the main screen after a successful payment.
    -   The `Promote Now` button now triggers the payment and boost initiation process.
    -   Data types in `EstimateResult.java` for budget and cost fields were changed from `int` to `double` for precision.

-   **New Analytics Dashboard (`AnalyticsDashboard.java`, `activity_analytics_dashboard.xml`)**:
    -   Created a new, detailed analytics screen accessible from the Professional Dashboard.
    -   The UI includes multiple chart types to visualize post performance:
        -   A **Line Chart** for "Total views" over time.
        -   A **Grouped Bar Chart** to break down "Media type" views by followers and non-followers.
        -   A **Donut (Pie) Chart** to show the "Viewer type" (followers vs. non-followers) percentage.
    -   The dashboard is populated with sample data and features horizontally scrollable cards for different metrics.

-   **API & Data Models**:
    -   Added `InitiateBoostRequest.java` to structure the data sent when boosting a post (e.g., budget, duration, audience, payment details).
    -   Added `BoostResponse.java` and related data models (`BoostData`) to handle the API response for the boost initiation.
    -   A new `initiateBoostPost` method was added to `UtilMethods.java` to handle the API call.

-   **UI/UX Enhancements**:
    -   Redesigned the `ProfessionalDashBoardPersonal` screen, introducing new "Analytics" and "Content" sections with a more compact and modern card-based layout.
    -   Several tabs on the professional dashboard screen were hidden to simplify the interface.
    -   Added new drawable resources, such as `bg_tab_selected.xml`, to improve the UI of the new analytics sections.
    -   Corrected a minor typo from "Whisma Plus" to "Wishma Plus" in the ad creation screen.

+++ app/src/main/java/com/infotech/wishmaplus/Api/Response/EstimateResult.java
@@ -3,9 +3,9 @@ package com.infotech.wishmaplus.Api.Response;
 public class EstimateResult {
 
     private String reach;
-    private int budget;
-    private int estimatedCost;
-    private int subTotal;
+    private double budget;
+    private double estimatedCost;
+    private double subTotal;
     private double gst;
     private double total;
     private double userBalance;
@@ -21,27 +21,27 @@ public class EstimateResult {
         this.reach = reach;
     }
 
-    public int getBudget() {
+    public double getBudget() {
         return budget;
     }
 
-    public void setBudget(int budget) {
+    public void setBudget(double budget) {
         this.budget = budget;
     }
 
-    public int getEstimatedCost() {
+    public double getEstimatedCost() {
         return estimatedCost;
     }
 
-    public void setEstimatedCost(int estimatedCost) {
+    public void setEstimatedCost(double estimatedCost) {
         this.estimatedCost = estimatedCost;
     }
 
-    public int getSubTotal() {
+    public double getSubTotal() {
         return subTotal;
     }
 
-    public void setSubTotal(int subTotal) {
+    public void setSubTotal(double subTotal) {
         this.subTotal = subTotal;
     }
